aaudio: account for hardware jitter in clock model
This will allow AAudio to track the DSP better when
the DSP has variable timing, such as when resampling.

Keep track of the maximum lateness.
Do not assume it is only one burst.
Account for the lateness on the input streams
so we don't try to read the data before the DSP
has written it.

Bug: 123643363
Test: record through iRig UA using Oboe Tester
Test: see bug for details
Change-Id: I49eb852c6d0324e8a26ee912da5108021ec113fe