Guard fd with unique_fd in C2OMXNode::emptyBuffer
The fd cloned near the beginning of C2OMXNode::emptyBuffer
should last until this function goes out of scope, but we
are seeing the fd becomes invalid during this process.
This causes a crash during Component::queue when we try
to clone from this fd again.

Use unique_fd to guard most of the life span of this fd in
hope to catch the offender.

bug: 133254412
test: camera recording; cts VideoEncoderTest;EncodeDecodeTest.

Change-Id: I4bd7398c069e58915303bbc0f58b1471c2755e31