Speed up SimpleDecodingSource
SimpleDecodingSource was waiting up to 5 milliseconds for each input
buffer, when it should not have been waiting at all, and instead
just check if there's an output buffer. This greatly increases
throughput for files with small buffers such as audio files, e.g.
a 6x speedup was observed for ogg and mp3 test streams.

Test: stagefright
Change-Id: If9fc91b823442926e8452d0e6cf95735b7248596