SoftVorbisDec: Fix isConfigured() in vorbis decoder plugin
isConfigured() in soft vorbis decoder omx plugin was using
mInputBufferCount to detect if the decoder was configured.
But mInputBufferCount was reset to 0 after flush.
This was prone to a race condition, when flush was called immediately
after start(), before codec could signal correct number of channels and
sampling rate.
isConfigured() is now fixed to check decoder state rather than
mInputBufferCount.

Bug: 137736256
Bug: 138213973
Test: atest android.media.cts.DecoderTest#testDecodeWithEOSOnLastBuffer

Change-Id: Ia1e6b88eca45b46fe20f36627f79cd767b73cc1a