Move ExtendedAudioBufferProvider and SingleStateQueue
Move ExtendedAudioBufferProvider.h to libaudioclient. This is
a more appropriate location because EABP extends AudioBufferProvider.

Move SingleStateQueue.h to libnbaio_mono. This is a more appropriate
location because SSQ is a non-blocking queue.

This allows to remove the dependency of libnbaio on libmedia
which is a good thing because libnbaio provides more low-level
abstractions than libmedia.

Also, replace a dependency of libnbaio on libbinder with
a dependency on libaudiohal header library.

Test: make
Change-Id: Ie48b523790cd8230695ec2e4710e50981b616289