Fix dst rect calculation
dst rect also need to be adjusted when the source
frame has crop. Otherwise the converter will refuse
to copy if src and dst size doesn't match.

This happens when the decoding is done by some
hw decoders where the frame size is large than
actual size and a crop is used.

test: MediaMetadataRetrieverTests;HeifWriterTest;
various heic files with or without tiles.

bug: 141418155
Change-Id: I2190f2a23ffb0785f14e4d5830987e898f2a36cb