aaudio: write if there is room
Go ahead and write the data to the buffer if there is room,
even if it exceeds the setBufferSizeInFrames().

There is no point of holding it in the application buffer.
By making the data available to the DSP we can reduce glitches
or reduce latency.

Because this offers better glitch protection, we can allow
setBufferSizeInFrames() to accept a buffer size of zero.
This will provide the lowest possible latency.

Test: OboeTester glitch test
Change-Id: Ie706867c5dd57c29b4393c806cd51ae09198873e