Add a comment on incorrect paths that are required for Q compatibility.
Searching /apex/com.android.runtime/${LIB} in the platform namespace makes
its libraries visible in the wrong namespace - they should be in a separate
"runtime" namespace to keep separation. Furthermore, in R the directory
only contains Bionic internal dependencies that no external package should
depend upon.

However the directory cannot be removed from the search paths because this
linker config is in an APEX that needs to work on Q too, where it is used to
find libclang_rt.hwasan*.so. (The issue with the libraries being visible in
the wrong namespace applies to Q too, but it doesn't have any known
observable effects.)

Hence only add a comment to highlight that it should be cleaned up when the
code has branched and Q compatibility can be dropped.

Test: Build & boot
Bug: 140790209
Bug: 139408016
Change-Id: I2a5e6418977200c56d01f52c0d78dd820cc643a2