Remove cleanUpEffectsForIo.
When closing input/output, it is not needed to unregister effects.
The effects will be automatically unregistered when the audio session id
is released. If the effects are unregistered when closing input/output
and the audio session id is not released, audio policy manager will lose
the visibility of the effects. In that case, it will fail to move the
effects to new I/O when a track with the same audio session id is
attached to a new I/O.

Bug: 145197421
Test: repo steps from the bug
Change-Id: I1bee3a3ee79692c863aa3977df62346ac9dc8946