Replace the shared lib dependency to libmedia with header deps
libstagefright_mpeg2support actually doesn't link with libmedia. It uses
headers from the library. Replace the shared lib dependency to the
library with the header lib dependency to libmedia_headers and
libaudioclient_headers.

Note: This was erroneously reverted in a subsequent patch. Hence
reuploading the change with a new Change Id to fix the revert.

Bug: N/A
Test: m libstagefright_mpeg2support
Change-Id: I20bf45b2eeb853eba653b539b1e1804272038cdc