Don't re-export libbinder headers from libstagefright_foundation
The headers of libstagefright_foundation do not reference libbinder
headers. Therefore, the libbinder headers don't need to be re-exported.

As a good side effect, libbinder is no longer built for
com.android.media APEX where libstagefright_foundation is a member of.
Before this change, due to the re-export of headers, the media APEX
variant of libbinder had to be created.

Bug: 147671264
Test: m
Change-Id: Id16054593487f042fba6e46b114da52439e98e07