Set min_sdk_version: 29
to ensure that native modules in media apex are built against stubs of
version 29.

But, setting min_sdk_version to swcodec apex doesn't change anything yet
because the build system doesn't support build against "min_sdk_version"
for vendor variants. It's kind of a preperation for the future support
of the build system.

Bug: 145796956
Test: m
      look into build.ninja to see if libs built against
      stubs of version 29 (for example, libc/libdl/libm/etc)
Change-Id: Ic1f3dafa45797feb29043391d24b87c4eed6f310